Change encryption method to support smaller tags
This version uses AES-128 to encrypt the password locally, with the
16-byte key stored on the NFC tag. This makes the NFC payload small
enough to support MIFARE Ultralight tags.

The previous scheme used a one-time pad, which meant that the NFC
payload had to be at least as long as the password (and ideally much longer,
so that the size of the payload didn't give any information away about
the length of the password).